# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Worldwide Rapid Task For Money - Verification Automation Engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, paused, completed, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignList"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /campaigns/{id}/stats:
    get:
      operationId: getCampaignStats
      tags: [campaigns]
      summary: Get campaign statistics
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignStats"
  /campaigns/{id}/tasks:
    get:
      operationId: getCampaignTasks
      tags: [campaigns]
      summary: List tasks for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"

  # Tasks
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
...
    Campaign:
      type: object
      required: [id, clientId, title, description, taskType, platform, status, budget, spent, workersNeeded, workersCompleted, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        description:
          type: string
        taskType:
          type: string
          enum: [youtube_watch, youtube_like, youtube_comment, youtube_subscribe, instagram_follow, instagram_like, instagram_comment, tiktok_follow, tiktok_like, tiktok_comment, twitter_follow, twitter_like, twitter_retweet, website_visit, website_signup, app_install, app_test, form_fill, content_review, data_collection]
        platform:
          type: string
        targetUrl:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, cancelled]
        budget:
          type: number
        spent:
          type: number
        rewardPerTask:
          type: number
        workersNeeded:
          type: integer
        workersCompleted:
          type: integer
        targetCountries:
          type: array
          items:
            type: string
        targetLanguages:
          type: array
          items:
            type: string
        deadline:
          type: ["string", "null"]
          format: date-time
        instructions:
          type: string
        proofRequirements:
          type: array
          items:
            type: string
        automationRate:
          type: number
        createdAt:
          type: string
          format: date-time
    CampaignList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Campaign"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CampaignInput:
      type: object
      required: [title, description, taskType, platform, budget, rewardPerTask, workersNeeded, instructions]
      properties:
        title:
          type: string
        description:
          type: string
        taskType:
          type: string
        platform:
          type: string
        targetUrl:
          type: string
        budget:
          type: number
        rewardPerTask:
          type: number
        workersNeeded:
          type: integer
        targetCountries:
          type: array
          items:
            type: string
        targetLanguages:
          type: array
          items:
            type: string
        deadline:
          type: string
          format: date-time
        instructions:
          type: string
        proofRequirements:
          type: array
          items:
            type: string
    CampaignUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, active, paused, cancelled]
        budget:
          type: number
        instructions:
          type: string
        deadline:
          type: string
          format: date-time
    CampaignStats:
      type: object
      required: [campaignId, totalAssignments, approved, rejected, pending, automationRate, avgConfidenceScore, totalPaid, avgCompletionTime]
      properties:
        campaignId:
          type: integer
        totalAssignments:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        pending:
          type: integer
        manualReview:
          type: integer
        automationRate:
          type: number
        avgConfidenceScore:
          type: number
        totalPaid:
          type: number
        avgCompletionTime:
          type: number
        byCountry:
          type: array
          items:
            $ref: "#/components/schemas/CountryBreakdown"
    CountryBreakdown:
      type: object
      required: [country, count]
      properties:
        country:
          type: string
        count:
          type: integer

    # Tasks
    Task:
      type: object
      required: [id, campaignId, title, taskType, platform, reward, status, createdAt]
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        title:
          type: string
        taskType:
          type: string
        platform:
          type: string
        targetUrl:
          type: ["string", "null"]
        reward:
          type: number
        status:
          type: string
          enum: [available, in_progress, completed, expired, cancelled]
        instructions:
          type: string
        proofRequirements:
          type: array
          items:
            type: string
        estimatedMinutes:
          type: integer
        deadline:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        campaign:
          $ref: "#/components/schemas/Campaign"
...