# backend/src/api-zod/generated/types/campaign.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Worldwide Rapid Task For Money - Verification Automation Engine
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignStatus } from './campaignStatus';
import type { CampaignTaskType } from './campaignTaskType';

export interface Campaign {
  id: number;
  clientId: number;
  title: string;
  description: string;
  taskType: CampaignTaskType;
  platform: string;
  targetUrl?: string;
  status: CampaignStatus;
  budget: number;
  spent: number;
  rewardPerTask?: number;
  workersNeeded: number;
  workersCompleted: number;
  targetCountries?: string[];
  targetLanguages?: string[];
  /** @nullable */
  deadline?: Date | null;
  instructions?: string;
  proofRequirements?: string[];
  automationRate?: number;
  createdAt: Date;
}